### 2023-08-18 #5764 @oliviaguest
Yes, it's likely you missed this:

> Your paper must not focus on new research results accomplished with the software.

https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements
