### 2023-08-29 #5764 @Nikoleta-v3
👋🏻 @hanneoberman, @EugeneHao, @josiahparry, @seabbs would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "verdata: An R package for analyzing data from the Truth Commission in Colombia". You can find more information at the top of this Github issue 🆙
