### 2023-09-11 #5764 @Nikoleta-v3
👋🏻 @bcjaeger @jamesmbaazam @y1my1 would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "verdata: An R package for analyzing data from the Truth Commission in Colombia". You can find more information at the top of this Github issue 🆙
