### 2023-09-11 #5764 @bcjaeger
@Nikoleta-v3 Hello! The verdata R package looks very interesting. If needed, I'd be happy to review and can start on September 25.
