### 2023-09-13 #5844 @editorialbot
[REVIEW]: verdata: An R package for analyzing data from the Truth Commission in Colombia
**Submitting author:** @thegargiulian (<a href="http://orcid.org/0000-0003-1870-8990">Maria Gargiulo</a>)
**Repository:** https://github.com/HRDAG/verdata
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v.0.9.1
