### 2023-09-13 #5764 @hanneoberman
> 👋🏻 @hanneoberman, @EugeneHao, @JosiahParry, @seabbs would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission I'd like you to review is titled: "verdata: An R package for analyzing data from the Truth Commission in Colombia". You can find more information at the top of this Github issue 🆙
>
> Please let me know if you're available 😄 Thank you!

Sorry for not responding sooner! I haven't been available, but I'd like to review other `R` packages in the future.
