### 2023-10-12 #5844 @JosiahParry
@thegargiulian please add a plain text license to the repository.
