### 2023-10-12 9c5af3d
adding plain text license, closes #19 and removing extdata files from .gitignore for #21
