### 2023-10-12 #5844 @thegargiulian
Thank you, @JosiahParry! The issues you opened are helpful and I'll get working on them as soon as @jamesmbaazam's comments are in.

For right now I've added a plain text license and fixed the `extdata` issue  (which we were still gitignoring from an earlier development phase).

Thanks also for your comment about making the examples into a quarto or bookdown book—that's a great suggestion for future work in that repo!
