### 2023-10-14 HRDAG/verdata#26 @jamesmbaazam
Move part of package description to README
In my opinion, the second sentence onwards might be better placed in the README as it seems to get lost in the package DESCRIPTION file.

https://github.com/HRDAG/verdata/blob/9c5af3d2bd72476a9c8518bfd85586e2a09564a6/DESCRIPTION#L20

_This issue is in relation to the JOSS review here https://github.com/openjournals/joss-reviews/issues/5844_
