### 2023-10-14 HRDAG/verdata#27 @jamesmbaazam
Installation instructions: Check if `devtools` is installed
Under the [installation instructions](https://github.com/HRDAG/verdata#instalaci%C3%B3n), it might be better to replace
`install.packages("devtools")` with `if (!require("devtools")) install.packages("devtools")` to only install `devtools` if it's not already installed. This could reduce unnecessary re-installations.

_This issue is with reference to the JOSS review here https://github.com/openjournals/joss-reviews/issues/5844_
