### 2023-10-14 #5844 @jamesmbaazam
@thegargiulian In the software paper, could you please provide evidence for lines 43-44?
