### 2023-10-14 #5844 @jamesmbaazam
@thegargiulian I really like the last but one paragraph in the Software paper. It is a major selling point and I would suggest weaving a summary of that into the "Summary" section. I think the essence I'm getting from that paragraph is that the package provides ease of use of the data, especially for researchers who are unfamiliar with using imputed datasets.
