### 2023-10-14 HRDAG/verdata#29 @jamesmbaazam
Provide contribution instructions
In the README, could you please provide information on how to contribute to this package and how to report issues, bugs, errors, and other problems with the package?

_This issue is with reference to the JOSS review here https://github.com/openjournals/joss-reviews/issues/5844_
