### 2023-10-14 HRDAG/verdata#30 @jamesmbaazam
Set up continuous integration
Could you set up continuous integration, for example with GitHub Actions, to check that the tests and infrastructure are working correctly? The package {usethis} provides quick functionalities to get started. See [here](https://usethis.r-lib.org/reference/github_actions.html).

_This issue is with reference to the JOSS review here https://github.com/openjournals/joss-reviews/issues/5844_
