### 2023-10-14 #5844 @jamesmbaazam
@thegargiulian The Software paper is very clear and well-written. However, it would be nice to provide a brief overview of related software/packages if any exist. If there is none, it might be helpful to state that by positioning {verdata} in the current ecosystem.
