### 2023-10-15 #5844 @thegargiulian
Thank you @jamesmbaazam for these helpful issues and comments on the paper!

I'll get working on everything you and @JosiahParry flagged ASAP and will hopefully have everything addressed in the next 1-2 weeks.
