### 2023-10-15 7606d53
add CI with usethis; #30
