+++ inst/docs/README-en.md
@@ -1,5 +1,7 @@
 Haga clic [aquí](https://github.com/HRDAG/verdata/blob/main/README.md) para instrucciones en español.
 
+[![R-CMD-check](https://github.com/HRDAG/verdata/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/HRDAG/verdata/actions/workflows/R-CMD-check.yaml)
+
 <div class="columns">
 
 <div class="column" width="40%">
