### 2023-10-15 HRDAG/verdata#30 @thegargiulian
This is the only issue I'm addressing without a pull request.

I've added CI using `usethis` as you've suggested and, after removing the `extdata` directory from `.gitignore`, everything is running as expected. This also addresses the points raised in #21.

_This issue is with reference to the JOSS review here https://github.com/openjournals/joss-reviews/issues/5844_
