### 2023-10-16 HRDAG/verdata#30 @jamesmbaazam
Thanks @thegargiulian, For the current `rcmdcheck` workflow, may I suggest you rather run `usethis::use_github_action("check-standard")` so that it checks on all the standard platforms?

Additionally, could you please add the test coverage workflow as well by running `usethis::use_github_action("test-coverage")`?

If you decide to set up the package website as suggested in #28 (which I strongly recommend), you could also [add the `pkg-down` workflow](https://github.com/r-lib/actions/tree/v2/examples#build-pkgdown-site) to automatically deploy the website whenever you make changes.
