### 2023-10-16 8969b35
adding workflows for check-standard and test-coverage #30
