### 2023-10-16 8c71997
updating estimation test to account for differences in windows test environment
