+++ inst/docs/README-en.md
@@ -1,6 +1,9 @@
 Haga clic [aquí](https://github.com/HRDAG/verdata/blob/main/README.md) para instrucciones en español.
 
+ <!-- badges: start -->
 [![R-CMD-check](https://github.com/HRDAG/verdata/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/HRDAG/verdata/actions/workflows/R-CMD-check.yaml)
+  [![Codecov test coverage](https://codecov.io/gh/HRDAG/verdata/branch/main/graph/badge.svg)](https://app.codecov.io/gh/HRDAG/verdata?branch=main)
+  <!-- badges: end -->
 
 <div class="columns">
 
