+++ DESCRIPTION
@@ -27,6 +27,7 @@ RoxygenNote: 7.2.3
 Depends:
     R (>= 3.5)
 Suggests:
+    covr,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 Imports:
