### 2023-10-16 4a42a86
swapping rcmdchk reference for badge
