### 2023-10-16 HRDAG/verdata#30 @thegargiulian
Thanks @jamesmbaazam!

I've replaced the `rcmdcheck` workflow with  `usethis::use_github_action("check-standard")` and setup the test coverage workflow using codecov and `usethis::use_github_action("test-coverage")`.
