### 2023-10-16 2f5c496
fixing title for get_valid_sources #23
