### 2023-10-21 18d21d7
fixing titles for run_lcmcr, proportions_imputed, and proportions_observed #25
