### 2023-10-23 f6852ce
add data dictionaries as R data objects #18
