+++ R/diccionario_vars_adicional.R
@@ -1,0 +1,24 @@
+# ============================================
+# Authors:     MG
+# Maintainers: MG
+# Copyright:   2023, HRDAG, GPL v2 or later
+# ============================================
+
+#' Variables adicionales que pueden ser útiles para analizar los datos.
+#' 
+#' @docType data
+#' 
+#' @usage data(diccionario_vars_adicional)
+#' @format Un data frame con 11 filas y 4 variables
+#' \describe{
+#'   \item{nombre_variable}{nombre de la variable}
+#'   \item{tipo}{tipo de la variable: caracter, numérico, lógico}
+#'   \item{detalle_variable}{explicación detallada de la variable}
+#'   \item{categorias_variable}{valores posibles de la variable}
+#' }
+#'
+#' @source proyecto conjunto JEP-CEV-HRDAG
+#'
+"diccionario_vars_adicional"
+
+# done.
