+++ R/diccionario_replicas.R
@@ -1,0 +1,25 @@
+# ============================================
+# Authors:     MG
+# Maintainers: MG
+# Copyright:   2023, HRDAG, GPL v2 or later
+# ============================================
+
+#' Diccionario de datos para las variables que aparecen en los archivos de las
+#' réplicas. 
+#' 
+#' @docType data
+#' 
+#' @usage data(diccionario_replicas)
+#' @format Un data frame con 55 filas y 4 variables
+#' \describe{
+#'   \item{nombre_variable}{nombre de la variable}
+#'   \item{tipo}{tipo de la variable: caracter, numérico, lógico}
+#'   \item{detalle_variable}{explicación detallada de la variable}
+#'   \item{categorias_variable}{valores posibles de la variable}
+#' }
+#'
+#' @source proyecto conjunto JEP-CEV-HRDAG
+#'
+"diccionario_replicas"
+
+# done.
