+++ man/diccionario_replicas.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/diccionario_replicas.R
+\docType{data}
+\name{diccionario_replicas}
+\alias{diccionario_replicas}
+\title{Diccionario de datos para las variables que aparecen en los archivos de las
+réplicas.}
+\format{
+Un data frame con 55 filas y 4 variables
+\describe{
+\item{nombre_variable}{nombre de la variable}
+\item{tipo}{tipo de la variable: caracter, numérico, lógico}
+\item{detalle_variable}{explicación detallada de la variable}
+\item{categorias_variable}{valores posibles de la variable}
+}
+}
+\source{
+proyecto conjunto JEP-CEV-HRDAG
+}
+\usage{
+data(diccionario_replicas)
+}
+\description{
+Diccionario de datos para las variables que aparecen en los archivos de las
+réplicas.
+}
+\keyword{datasets}
