+++ R/diccionario_replicas.R
@@ -10,7 +10,7 @@
 #' @docType data
 #' 
 #' @usage data(diccionario_replicas)
-#' @format Un data frame con 55 filas y 4 variables
+#' @format Un data frame con 55 filas y 4 variables.
 #' \describe{
 #'   \item{nombre_variable}{nombre de la variable}
 #'   \item{tipo}{tipo de la variable: caracter, numérico, lógico}
@@ -18,7 +18,7 @@
 #'   \item{categorias_variable}{valores posibles de la variable}
 #' }
 #'
-#' @source proyecto conjunto JEP-CEV-HRDAG
+#' @source Proyecto conjunto JEP-CEV-HRDAG.
 #'
 "diccionario_replicas"
 
