### 2023-10-23 c0e74c1
removing test that is unreliable on windows test builds
