### 2023-10-23 c1acd1c
only install devtools if necessary; #27
