### 2023-10-23 86c3f7c
clarify GSL installation language to not confuse users by mentioning LCMCR when not necessary #17
