### 2023-10-23 e2e6c17
add information about contribution instructions #29
