### 2023-10-23 0f9c64d
adding information from package description to README #26
