### 2023-10-23 d81e1ff
removing extra info from description file #26
