### 2023-10-25 0a12fa8
translate changes from Spanish version of README to English version
