### 2023-10-25 #5844 @thegargiulian
Hi @JosiahParry and @jamesmbaazam,

Thank you both again for your helpful issues and comments!

The majority of the issues you opened were addressed in https://github.com/HRDAG/verdata/pull/31, https://github.com/HRDAG/verdata/pull/32, and https://github.com/HRDAG/verdata/pull/33. I added a plain text `LICENSE` file and the infrastructure to do automatic testing using GitHub actions directly to `main`, so there is no pull request associated with those changes.

@jamesmbaazam, I have implemented your suggested changes to the manuscript and you will find an updated version of the manuscript in the comment above.

Thank you again and please let me know if there are any additional changes you think are necessary for the software or the paper!
