### 2023-11-09 #5844 @Nikoleta-v3
Hey @jamesmbaazam and @JosiahParry 👋🏻 😄  I was wondering if you had any updates regarding @thegargiulian's response: https://github.com/openjournals/joss-reviews/issues/5844#issuecomment-1779450867
