### 2023-11-10 #5844 @JosiahParry
@thegargiulian all boxes ticked on my end. I also feel like a complete fool. I only _just_ realized that `verdata` is a play on "verdad" / the truth commission! 🤦🏼

@Nikoleta-v3 totally unrelated, but is there a way to associate my ORC ID with my reviewer username?
