### 2023-11-28 #5844 @Nikoleta-v3
Hello @thegargiulian, thank you very much for your patience! My apologies for the delay. I ended up catching COVID and then went on a break. Anyway, my apologies again!

Thank you for your submission. I believe it's a fantastic package, and I am very happy to see how productive the review process was!

I have some questions and comments:

**Regarding the documentation:**

Could you please include some details on how a user could run the tests?

In the contribution guidelines, it is stated that comments and suggestions are very welcome, but it is possible that someone would like to contribute some code. If you accept code contributions, then you should mention, in the contribution guidelines, that they should open a pull request and that there is CI setup that will run tests upon a pull request. One could also test the package locally by following "these" instructions.

If you decide to include this, I can see why including all this information in the readme might be a bit much. If you would like to include more information, consider having a file. See the [example](https://github.com/FRBCesab/chessboard/blob/0e8081a9465b876fbf27619fe1b54377e5ca1bcc/CONTRIBUTING.md). Note that this is a very detailed contribution file.

**Regarding the examples:**

Unfortunately, I can't run the examples since my Spanish is nonexistent, but I am very much looking forward to the English version. One question I have is how one should follow these examples. Is the website hosted somewhere? Or should one clone the repository, build the website locally, or open the HTML files? I am really sorry if you are already discussing this somewhere, and I missed it!

**Regarding the paper:**

Line 24 should be `create to support`.

In `paper.bib` reference `vanbuuren2011`, could you change the title to be  `title={mice: Multivariate imputation by chained equations in {R}}`. This will make sure that in the paper, in the list of references, the R will appear as a capital letter.
