### 2023-12-08 5ec7a2f
update readme with information about CI and testing
