### 2023-12-11 e03b66d
upping version for JOSS paper
