### 2023-12-11 #5844 @thegargiulian
Thanks, @Nikoleta-v3!

I've tagged the repository as version 0.9.1 and you'll find the package on Zenodo at https://doi.org/10.5281/zenodo.10359636.

Thank you and let me know if you need anything else from my end!
