### 2023-12-12 #5844 @thegargiulian
@Nikoleta-v3, so sorry! I just found a typo in footnote 2. Is it possible to fix it?
