### 2024-01-03 #5844 @thegargiulian
Hey @Nikoleta-v3, happy new year!

Just wondering if you had  status update on this—I think we're just waiting on the EICs, but let me know if you need anything from our end. Thank you!
