### 2024-01-06 #5844 @oliviaguest
Huge thanks to the reviewers @jamesmbaazam, @JosiahParry and editor @Nikoleta-v3! ✨  JOSS appreciates your work and effort. ✨ Also, big congratulations to the authors @thegargiulian! 🥳 🍾
